WIP revamp storage

+++ contracts/libraries/ShareMath.sol
@@ -13,7 +13,7 @@ library ShareMath {
     function underlyingToShares(
         uint256 underlyingAmount,
         uint256 pps,
-        uint8 decimals
+        uint256 decimals
     ) internal pure returns (uint104) {
         // If this throws, it means that vault's roundPricePerShare[currentRound] has not been set yet
         // which should never happen.
@@ -21,7 +21,7 @@ library ShareMath {
         require(pps > PLACEHOLDER_UINT, "Invalid pps");
 
         uint256 shares =
-            uint256(underlyingAmount).mul(10**uint256(decimals)).div(pps);
+            uint256(underlyingAmount).mul(10**decimals).div(pps);
         assertUint104(shares);
 
         return uint104(shares);
@@ -30,7 +30,7 @@ library ShareMath {
     function sharesToUnderlying(
         uint256 shares,
         uint256 pps,
-        uint8 decimals
+        uint256 decimals
     ) internal pure returns (uint256) {
         // If this throws, it means that vault's roundPricePerShare[currentRound] has not been set yet
         // which should never happen.
@@ -38,7 +38,7 @@ library ShareMath {
         require(pps > PLACEHOLDER_UINT, "Invalid pps");
 
         uint256 underlyingAmount =
-            uint256(shares).mul(pps).div(10**uint256(decimals));
+            uint256(shares).mul(pps).div(10**decimals);
         assertUint104(shares);
 
         return underlyingAmount;
@@ -56,7 +56,7 @@ library ShareMath {
         Vault.DepositReceipt memory depositReceipt,
         uint256 currentRound,
         uint256 pps,
-        uint8 decimals
+        uint256 decimals
     ) internal pure returns (uint128 unredeemedShares) {
         if (depositReceipt.round > 0 && depositReceipt.round < currentRound) {
             uint256 sharesFromRound =
